Created humidity schemas

+++ app/models/humidity_records.py
@@ -10,5 +10,5 @@ humidity_records_table = sqlalchemy.Table(
     sqlalchemy.Column("id", sqlalchemy.Integer, primary_key=True),
     sqlalchemy.Column("location_id", sqlalchemy.ForeignKey(locations_table.c.id)),
     sqlalchemy.Column("date", sqlalchemy.DateTime()),
-    sqlalchemy.Column("humidity", sqlalchemy.Float)
+    sqlalchemy.Column("humidity", sqlalchemy.Integer)
 )
